Added april tags to pose calculation using a kalman filter. The accuracy of the calculation is still untested.

+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/Helpers/Utils.java
@@ -21,7 +21,7 @@ public class Utils {
      * @param time   float time to wait in seconds
      * @param opMode used to check if opMode is still active
      */
-    public static void sleep(float time, LinearOpMode opMode) {
+    public static void sleep(double time, LinearOpMode opMode) {
         ElapsedTime timer = new ElapsedTime();
         timer.reset();
         while (timer.seconds() < time && opMode.opModeIsActive()) {
